// src/app/heroes/interfaces/hero.interface.ts
// Generated by https://quicktype.io

export interface Hero {
  id:               string;
  superhero:        string;
  publisher:        Publisher;
  alter_ego:        string;
  first_appearance: string;
  characters:       string;
  //Creo nueva propiedad para el pipe HeroImagePipe
  alt_img?:         string;
}

export enum Publisher {
  DCComics = "DC Comics",
  MarvelComics = "Marvel Comics",
}